+++ src/event/grid_destroy.rs
@@ -4,7 +4,7 @@ use rmpv::Value;
 /// Grid will not be used anymore and the UI can free any data associated with it.
 #[derive(Debug, Clone)]
 pub struct GridDestroy {
-    pub grid: u64,
+    pub grid: u32,
 }
 
 impl Parse for GridDestroy {
